pyproject.toml
[project]
name = "secret-hitler-bench"
version = "0.1.0"
description = "LLM social deduction benchmark using Secret Hitler"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "playwright>=1.40",
    "websockets>=12.0",
]

[project.scripts]
sh-bench = "main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
